backend/requirement.txt
fastapi
uvicorn
openai
requests
python-jose
passlib[bcrypt]
sqlalchemy
asyncpg
databases
